[refactor] Extract cmd/list_render.go
Pure move — the table/JSON rendering helpers come out of list.go into
a sibling file so the command file stays focused on the flag-to-pipeline
flow. No signature or behavior change yet; that lands when list.go is
wired through operations.ListWorktrees.

Moved: listRenderEmpty, listRenderJSON, listRenderTable, listHeader,
listRow, formatPRCell, formatCICell.

Refs #132
